backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai>=0.3.1,<0.4.0
pydantic==2.5.0
cors==1.0.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
